Restructure ThermoFisher thermocycler into multi-file package
Split pylabrobot/thermo_fisher/thermocycler.py into:
- thermocyclers/driver.py — raw SCPI IO, auth, block discovery, power
- thermocyclers/block_backend.py — per-block temp control via send_command
- thermocyclers/lid_backend.py — per-block lid temp control via send_command
- thermocyclers/thermocycling_backend.py — protocol execution, run progress
- thermocyclers/thermocycler.py — device class
- thermocyclers/utils.py — XML/SCPI helpers, RunProgress
- thermocyclers/proflex.py, atc.py — factory functions

Backends call driver.send_command() directly — driver is pure IO layer.
Backward-compat shim at thermocycler.py re-exports from package.
All 3 legacy proflex tests pass.
