+++ pylabrobot/legacy/thermocycling/thermo_fisher/thermo_fisher_thermocycler.py
@@ -5,8 +5,11 @@ All real SCPI logic lives in :class:`ThermoFisherThermocyclerDriver`
 that :class:`ATCBackend`, :class:`ProflexBackend` and their tests continue to work unchanged.
 """
 
+import asyncio
+import re
 from abc import ABCMeta
-from typing import Dict, List, Optional
+from base64 import b64decode
+from typing import Dict, List, Optional, cast
 
 from pylabrobot.legacy.thermocycling.backend import ThermocyclerBackend
 from pylabrobot.legacy.thermocycling.standard import (
@@ -16,7 +19,7 @@ from pylabrobot.legacy.thermocycling.standard import (
   Step,
   protocol_to_new,
 )
-from pylabrobot.thermo_fisher.thermocycler import (
+from pylabrobot.thermo_fisher.thermocyclers import (
   RunProgress,
   ThermoFisherThermocyclerDriver,
   _gen_protocol_data,
@@ -142,11 +145,27 @@ class ThermoFisherThermocyclerBackend(ThermocyclerBackend, metaclass=ABCMeta):
     self, temperature: List[float], block_id: Optional[int] = None, rate: float = 100
   ):
     assert block_id is not None, "block_id must be specified"
-    await self._driver.set_block_temperature(temperature=temperature, block_id=block_id, rate=rate)
+    if block_id not in self._driver.available_blocks:
+      raise ValueError(f"Block {block_id} not available")
+    res = await self._driver.send_command(
+      {"cmd": f"TBC{block_id + 1}:RAMP", "params": {"rate": rate}, "args": temperature},
+      response_timeout=60,
+    )
+    if self._driver._parse_scpi_response(res)["status"] != "OK":
+      raise ValueError("Failed to ramp block temperature")
 
   async def set_lid_temperature(self, temperature: List[float], block_id: Optional[int] = None):
     assert block_id is not None, "block_id must be specified"
-    await self._driver.set_lid_temperature(temperature=temperature, block_id=block_id)
+    assert len(set(temperature)) == 1, "Lid temperature must be the same for all zones"
+    target_temp = temperature[0]
+    if block_id not in self._driver.available_blocks:
+      raise ValueError(f"Block {block_id} not available")
+    res = await self._driver.send_command(
+      {"cmd": f"TBC{block_id + 1}:CoverRAMP", "params": {}, "args": [target_temp]},
+      response_timeout=60,
+    )
+    if self._driver._parse_scpi_response(res)["status"] != "OK":
+      raise ValueError("Failed to ramp cover temperature")
 
   async def deactivate_lid(self, block_id: Optional[int] = None):
     assert block_id is not None, "block_id must be specified"
@@ -157,11 +176,13 @@ class ThermoFisherThermocyclerBackend(ThermocyclerBackend, metaclass=ABCMeta):
     await self._driver.deactivate_block(block_id=block_id)
 
   async def get_block_current_temperature(self, block_id=1) -> List[float]:
-    return await self._driver.get_block_current_temperature(block_id=block_id)
+    res = await self._driver.send_command({"cmd": f"TBC{block_id + 1}:TBC:BlockTemperatures?"})
+    return cast(List[float], self._driver._parse_scpi_response(res)["args"])
 
   async def get_lid_current_temperature(self, block_id: Optional[int] = None) -> List[float]:
     assert block_id is not None, "block_id must be specified"
-    return await self._driver.get_lid_current_temperature(block_id=block_id)
+    res = await self._driver.send_command({"cmd": f"TBC{block_id + 1}:TBC:CoverTemperatures?"})
+    return cast(List[float], self._driver._parse_scpi_response(res)["args"])
 
   async def run_protocol(
     self,
@@ -182,7 +203,7 @@ class ThermoFisherThermocyclerBackend(ThermocyclerBackend, metaclass=ABCMeta):
       if isinstance(stage, Step):
         protocol.stages[i] = Stage(steps=[stage], repeats=1)
     new_protocol = protocol_to_new(protocol)
-    await self._driver.run_protocol(
+    await self._run_protocol_impl(
       protocol=new_protocol,
       block_max_volume=block_max_volume,
       block_id=block_id,
@@ -195,19 +216,169 @@ class ThermoFisherThermocyclerBackend(ThermocyclerBackend, metaclass=ABCMeta):
       stage_name_prefixes=stage_name_prefixes,
     )
 
+  async def _run_protocol_impl(
+    self,
+    protocol,
+    block_max_volume: float,
+    block_id: int,
+    run_name: str = "testrun",
+    user: str = "Admin",
+    run_mode: str = "Fast",
+    cover_temp: float = 105,
+    cover_enabled: bool = True,
+    protocol_name: str = "PCR_Protocol",
+    stage_name_prefixes: Optional[List[str]] = None,
+  ):
+    from pylabrobot.capabilities.thermocycling import Stage as NewStage, Step as NewStep
+
+    if await self.check_run_exists(run_name):
+      pass  # run already exists
+    else:
+      await self.create_run(run_name)
+
+    # wrap all Steps in Stage objects where necessary
+    for i, stage in enumerate(protocol.stages):
+      if isinstance(stage, NewStep):
+        protocol.stages[i] = NewStage(steps=[stage], repeats=1)
+
+    for stage in protocol.stages:
+      for step in stage.steps:
+        if len(step.temperature) != self._driver.num_temp_zones:
+          raise ValueError(
+            f"Each step in the protocol must have a list of temperatures "
+            f"of length {self._driver.num_temp_zones}. "
+            f"Step temperatures: {step.temperature} (length {len(step.temperature)})"
+          )
+
+    stage_name_prefixes = stage_name_prefixes or ["Stage_" for _ in range(len(protocol.stages))]
+
+    # write run info files
+    xmlfile, tmpfile = _generate_run_info_files(
+      protocol=protocol,
+      block_id=block_id,
+      sample_volume=block_max_volume,
+      run_mode=run_mode,
+      protocol_name=protocol_name,
+      cover_temp=cover_temp,
+      cover_enabled=cover_enabled,
+      user_name="LifeTechnologies",
+    )
+    await self._driver._write_file(f"runs:{run_name}/{protocol_name}.method", xmlfile)
+    await self._driver._write_file(f"runs:{run_name}/{run_name}.tmp", tmpfile)
+
+    # load and run protocol via SCPI
+    load_res = await self._driver.send_command(
+      data=_gen_protocol_data(
+        protocol=protocol,
+        block_id=block_id,
+        sample_volume=block_max_volume,
+        run_mode=run_mode,
+        cover_temp=cover_temp,
+        cover_enabled=cover_enabled,
+        protocol_name=protocol_name,
+        stage_name_prefixes=stage_name_prefixes,
+      ),
+      response_timeout=5,
+      read_once=False,
+    )
+    if self._driver._parse_scpi_response(load_res)["status"] != "OK":
+      raise ValueError("Protocol failed to load")
+
+    start_res = await self._driver.send_command(
+      {
+        "cmd": f"TBC{block_id + 1}:RunProtocol",
+        "params": {
+          "User": user,
+          "CoverTemperature": cover_temp,
+          "CoverEnabled": "On" if cover_enabled else "Off",
+        },
+        "args": [protocol_name, run_name],
+      },
+      response_timeout=2,
+      read_once=False,
+    )
+
+    if self._driver._parse_scpi_response(start_res)["status"] != "NEXT":
+      raise ValueError("Protocol failed to start")
+
+    total_time = await self.get_estimated_run_time(block_id=block_id)
+    total_time = float(total_time)
+    self._driver.current_runs[block_id] = run_name
+
   async def get_run_info(self, protocol: Protocol, block_id: int) -> RunProgress:
-    return await self._driver.get_run_info(protocol=protocol_to_new(protocol), block_id=block_id)
+    new_protocol = protocol_to_new(protocol)
+    progress = await self._get_run_progress(block_id=block_id)
+    run_name = await self.get_run_name(block_id=block_id)
+    if not progress:
+      return RunProgress(
+        running=False,
+        stage="completed",
+        elapsed_time=await self.get_elapsed_run_time_from_log(run_name=run_name),
+        remaining_time=0,
+      )
+
+    if progress["RunTitle"] == "-":
+      await self._driver._read_response(timeout=5)
+      return RunProgress(
+        running=False,
+        stage="completed",
+        elapsed_time=await self.get_elapsed_run_time_from_log(run_name=run_name),
+        remaining_time=0,
+      )
+
+    if progress["Stage"] == "POSTRun":
+      return RunProgress(
+        running=True,
+        stage="POSTRun",
+        elapsed_time=await self.get_elapsed_run_time_from_log(run_name=run_name),
+        remaining_time=0,
+      )
+
+    time_elapsed = await self.get_elapsed_run_time(block_id=block_id)
+    remaining_time = await self.get_remaining_run_time(block_id=block_id)
+
+    if progress["Stage"] != "-" and progress["Step"] != "-":
+      current_step = new_protocol.stages[int(progress["Stage"]) - 1].steps[
+        int(progress["Step"]) - 1
+      ]
+      if current_step.hold_seconds == float("inf"):
+        while True:
+          block_temps = await self.get_block_current_temperature(block_id=block_id)
+          target_temps = current_step.temperature
+          if all(
+            abs(float(block_temps[i]) - target_temps[i]) < 0.5 for i in range(len(block_temps))
+          ):
+            break
+          await asyncio.sleep(5)
+        return RunProgress(
+          running=False,
+          stage="infinite_hold",
+          elapsed_time=time_elapsed,
+          remaining_time=remaining_time,
+        )
+
+    return RunProgress(
+      running=True,
+      stage=progress["Stage"],
+      elapsed_time=time_elapsed,
+      remaining_time=remaining_time,
+    )
 
   async def abort_run(self, block_id: int):
     await self._driver.abort_run(block_id=block_id)
 
   async def continue_run(self, block_id: int):
-    await self._driver.continue_run(block_id=block_id)
+    for _ in range(3):
+      await asyncio.sleep(1)
+      res = await self._driver.send_command({"cmd": f"TBC{block_id + 1}:CONTinue"})
+      if self._driver._parse_scpi_response(res)["status"] != "OK":
+        raise ValueError("Failed to continue from indefinite hold")
 
   # -- convenience delegations -------------------------------------------------
 
   async def get_sample_temps(self, block_id=1) -> List[float]:
-    return await self._driver.get_sample_temps(block_id=block_id)
+    res = await self._driver.send_command({"cmd": f"TBC{block_id + 1}:TBC:SampleTemperatures?"})
+    return cast(List[float], self._driver._parse_scpi_response(res)["args"])
 
   async def get_nickname(self) -> str:
     return await self._driver.get_nickname()
@@ -216,10 +387,32 @@ class ThermoFisherThermocyclerBackend(ThermocyclerBackend, metaclass=ABCMeta):
     await self._driver.set_nickname(nickname)
 
   async def get_log_by_runname(self, run_name: str) -> str:
-    return await self._driver.get_log_by_runname(run_name)
+    res = await self._driver.send_command(
+      {"cmd": "FILe:READ?", "args": [f"RUNS:{run_name}/{run_name}.log"]},
+      response_timeout=5,
+      read_once=False,
+    )
+    if self._driver._parse_scpi_response(res)["status"] != "OK":
+      raise ValueError("Failed to get log")
+    res.replace("\n", "")
+    encoded_log_match = re.search(r"<quote>(.*?)</quote>", res, re.DOTALL)
+    if not encoded_log_match:
+      raise ValueError("Failed to parse log content")
+    encoded_log = encoded_log_match.group(1).strip()
+    log = b64decode(encoded_log).decode("utf-8")
+    return log
 
   async def get_elapsed_run_time_from_log(self, run_name: str) -> int:
-    return await self._driver.get_elapsed_run_time_from_log(run_name)
+    """Parse a log to find the elapsed run time in hh:mm:ss format and convert to total seconds."""
+    log = await self.get_log_by_runname(run_name)
+    elapsed_time_match = re.search(r"Run Time:\s*(\d+):(\d+):(\d+)", log)
+    if not elapsed_time_match:
+      raise ValueError("Failed to parse elapsed time from log. Expected hh:mm:ss format.")
+    hours = int(elapsed_time_match.group(1))
+    minutes = int(elapsed_time_match.group(2))
+    seconds = int(elapsed_time_match.group(3))
+    total_seconds = (hours * 3600) + (minutes * 60) + seconds
+    return total_seconds
 
   async def set_block_idle_temp(
     self, temp: float, block_id: int, control_enabled: bool = True
@@ -236,7 +429,14 @@ class ThermoFisherThermocyclerBackend(ThermocyclerBackend, metaclass=ABCMeta):
     )
 
   async def block_ramp_single_temp(self, target_temp: float, block_id: int, rate: float = 100):
-    await self._driver.block_ramp_single_temp(target_temp=target_temp, block_id=block_id, rate=rate)
+    if block_id not in self._driver.available_blocks:
+      raise ValueError(f"Block {block_id} not available")
+    res = await self._driver.send_command(
+      {"cmd": f"TBC{block_id + 1}:BlockRAMP", "params": {"rate": rate}, "args": [target_temp]},
+      response_timeout=60,
+    )
+    if self._driver._parse_scpi_response(res)["status"] != "OK":
+      raise ValueError("Failed to ramp block temperature")
 
   async def buzzer_on(self):
     await self._driver.buzzer_on()
@@ -254,33 +454,82 @@ class ThermoFisherThermocyclerBackend(ThermocyclerBackend, metaclass=ABCMeta):
     await self._driver.power_off()
 
   async def check_run_exists(self, run_name: str) -> bool:
-    return await self._driver.check_run_exists(run_name)
+    res = await self._driver.send_command(
+      {"cmd": "RUNS:EXISTS?", "args": [run_name], "params": {"type": "folders"}}
+    )
+    if self._driver._parse_scpi_response(res)["status"] != "OK":
+      raise ValueError("Failed to check if run exists")
+    return cast(str, self._driver._parse_scpi_response(res)["args"][1]) == "True"
 
   async def create_run(self, run_name: str):
-    return await self._driver.create_run(run_name)
+    res = await self._driver.send_command(
+      {"cmd": "RUNS:NEW", "args": [run_name]}, response_timeout=10
+    )
+    if self._driver._parse_scpi_response(res)["status"] != "OK":
+      raise ValueError("Failed to create run")
+    return self._driver._parse_scpi_response(res)["args"][0]
 
   async def get_run_name(self, block_id: int) -> str:
     return await self._driver.get_run_name(block_id=block_id)
 
   async def get_estimated_run_time(self, block_id: int):
-    return await self._driver.get_estimated_run_time(block_id=block_id)
+    res = await self._driver.send_command({"cmd": f"TBC{block_id + 1}:ESTimatedTime?"})
+    if self._driver._parse_scpi_response(res)["status"] != "OK":
+      raise ValueError("Failed to get estimated run time")
+    return self._driver._parse_scpi_response(res)["args"][0]
 
   async def get_elapsed_run_time(self, block_id: int):
-    return await self._driver.get_elapsed_run_time(block_id=block_id)
+    res = await self._driver.send_command({"cmd": f"TBC{block_id + 1}:ELAPsedTime?"})
+    if self._driver._parse_scpi_response(res)["status"] != "OK":
+      raise ValueError("Failed to get elapsed run time")
+    return int(self._driver._parse_scpi_response(res)["args"][0])
 
   async def get_remaining_run_time(self, block_id: int):
-    return await self._driver.get_remaining_run_time(block_id=block_id)
+    res = await self._driver.send_command({"cmd": f"TBC{block_id + 1}:REMainingTime?"})
+    if self._driver._parse_scpi_response(res)["status"] != "OK":
+      raise ValueError("Failed to get remaining run time")
+    return int(self._driver._parse_scpi_response(res)["args"][0])
 
   async def get_error(self, block_id):
     return await self._driver.get_error(block_id=block_id)
 
   async def get_current_cycle_index(self, block_id: Optional[int] = None) -> int:
     assert block_id is not None, "block_id must be specified"
-    return await self._driver.get_current_cycle_index(block_id=block_id)
+    progress = await self._get_run_progress(block_id=block_id)
+    if progress is None:
+      raise RuntimeError("No progress information available")
+    if progress["RunTitle"] == "-":
+      await self._driver._read_response(timeout=5)
+      raise RuntimeError("Protocol completed or not started")
+    if progress["Stage"] == "POSTRun":
+      raise RuntimeError("Protocol in POSTRun stage, no current cycle index")
+    if progress["Stage"] != "-" and progress["Step"] != "-":
+      return int(progress["Stage"]) - 1
+    raise RuntimeError("Current cycle index is not available, protocol may not be running")
 
   async def get_current_step_index(self, block_id: Optional[int] = None) -> int:
     assert block_id is not None, "block_id must be specified"
-    return await self._driver.get_current_step_index(block_id=block_id)
+    progress = await self._get_run_progress(block_id=block_id)
+    if progress is None:
+      raise RuntimeError("No progress information available")
+    if progress["RunTitle"] == "-":
+      await self._driver._read_response(timeout=5)
+      raise RuntimeError("Protocol completed or not started")
+    if progress["Stage"] == "POSTRun":
+      raise RuntimeError("Protocol in POSTRun stage, no current cycle index")
+    if progress["Stage"] != "-" and progress["Step"] != "-":
+      return int(progress["Step"]) - 1
+    raise RuntimeError("Current step index is not available, protocol may not be running")
+
+  async def _get_run_progress(self, block_id: int):
+    res = await self._driver.send_command({"cmd": f"TBC{block_id + 1}:RUNProgress?"})
+    parsed_res = self._driver._parse_scpi_response(res)
+    if parsed_res["status"] != "OK":
+      raise ValueError("Failed to get run status")
+    if parsed_res["cmd"] == f"TBC{block_id + 1}:RunProtocol":
+      await self._driver._read_response()
+      return False
+    return self._driver._parse_scpi_response(res)["params"]
 
   # -- stubs for abstract methods not implemented on this hardware -------------
