+++ pylabrobot/thermo_fisher/__init__.py
@@ -1,6 +1,7 @@
-from .atc import ATC
-from .proflex import ProFlexSingleBlock, ProFlexThreeBlock
-from .thermocycler import (
+from .thermocyclers import (
+  ATC,
+  ProFlexSingleBlock,
+  ProFlexThreeBlock,
   ThermoFisherBlockBackend,
   ThermoFisherLidBackend,
   ThermoFisherThermocycler,
